split user into separate app
