+++ nextcloudappstore/user/templates/user/delete-account.html
